AR-663: Removed POC from names and added tests
