+++ config/api_platform/playlist.yaml
@@ -62,70 +62,6 @@ resources:
               in: path
               required: true
 
-      putSlide:
-        method: PUT
-        path: /playlists/{id}/slide/{slideId}
-        controller: App\Controller\PlaylistSlidePutController
-        read: false
-        deserialize: false
-        validate: false
-        openapi_context:
-          requestBody:
-            content:
-              application/ld+json:
-                schema:
-                  type: object
-                  properties:
-                    weight:
-                      type: integer
-          operationId: put-v1-playlist-slide
-          description: Link slide resource to playlist resource.
-          summary: Link slide resource to playlist resource.
-          tags:
-            - Playlists
-          parameters:
-            - schema:
-                type: string
-                format: id
-                pattern: "^[A-Za-z0-9]{26}$"
-              name: id
-              in: path
-              required: true
-            - schema:
-                type: string
-                format: id
-                pattern: "^[A-Za-z0-9]{26}$"
-              name: slideId
-              in: path
-              required: true
-
-      deleteSlide:
-        method: DELETE
-        path: /playlists/{id}/slide/{slideId}
-        controller: App\Controller\PlaylistSlideDeleteController
-        read: false
-        openapi_context:
-          operationId: delete-v1-playlist-slide
-          description: Unlink slide resource from playlist resource.
-          summary: Unlink slide resource from playlist resource.
-          tags:
-            - Playlists
-          parameters:
-            - schema:
-                type: string
-                format: id
-                pattern: "^[A-Za-z0-9]{26}$"
-              name: id
-              in: path
-              required: true
-            - schema:
-                type: string
-                format: id
-                pattern: "^[A-Za-z0-9]{26}$"
-              name: slideId
-              in: path
-              required: true
-
     collectionOperations:
       get:
         openapi_context:
@@ -191,46 +127,6 @@ resources:
               in: query
               name: itemsPerPage
               description: The number of items per page
-
-      getPlaylistWithSlide:
-        method: GET
-        path: /slides/{id}/playlists
-        controller: App\Controller\PlaylistSlideGetController
-        read: false
-        openapi_context:
-          operationId: get-v1-playlist-with-slides
-          description: Retrieves a collection of Playlists resources.
-          summary: Retrieves a collection of Playlists resources base on Slide.
-          tags:
-            - Slides
-          parameters:
-            - schema:
-                type: string
-                format: id
-                pattern: "^[A-Za-z0-9]{26}$"
-              name: id
-              in: path
-              required: true
-            - schema:
-                type: integer
-                minimum: 0
-                format: int32
-                default: 1
-              in: query
-              name: page
-            - schema:
-                type: string
-                default: '10'
-              in: query
-              name: itemsPerPage
-              description: The number of items per page
-          responses:
-            '200':
-              description: OK
-              content:
-                application/ld+json:
-                  examples:
-              headers: { }
 
       post:
         openapi_context:
