+++ config/api_platform/playlist_slide.yaml
@@ -11,8 +11,8 @@ resources:
         controller: App\Controller\NotFoundAction
 
       delete:
-        path: /poc/playlist/{id}/slides/{slideId}
-        controller: App\Controller\PlaylistSlidePOCDeleteController
+        path: /playlists/{id}/slides/{slideId}
+        controller: App\Controller\PlaylistSlideDeleteController
         read: false
         deserialize: false
         validate: false
@@ -21,7 +21,7 @@ resources:
           summary: Delete a slide from a playlist.
           operationId: delete-v1-playlist-id
           tags:
-            - PlaylistSlide
+            - Playlists
           parameters:
             - schema:
                 type: string
@@ -39,8 +39,8 @@ resources:
               required: true
 
       put:
-        path: /poc/playlist/{id}/slides
-        controller: App\Controller\PlaylistSlidePOCPutController
+        path: /playlists/{id}/slides
+        controller: App\Controller\PlaylistSlidePutController
         read: false
         deserialize: false
         validate: false
@@ -49,7 +49,7 @@ resources:
           summary: Update the collection of slide on a playlist.
           operationId: put-v1-playlist-id
           tags:
-            - PlaylistSlide
+            - Playlists
           requestBody:
             content:
               application/ld+json:
@@ -84,8 +84,8 @@ resources:
 
     collectionOperations:
       get:
-        path: /poc/playlist/{id}/slides
-        controller: App\Controller\PlaylistSlidePOCGetController
+        path: /playlists/{id}/slides
+        controller: App\Controller\PlaylistSlideGetController
         read: false
         deserialize: false
         validate: false
@@ -94,7 +94,7 @@ resources:
           summary: Retrieves collection of weight Slide resources.
           operationId: get-v1-playlistslide-id
           tags:
-            - PlaylistSlide
+            - Playlists
           parameters:
             - schema:
                 type: string
